requirements.txt
django==4.0.4

django-redis==5.4.0
django-environ==0.9.0
django_cleanup==6.0.0
psycopg2-binary
django-tailwind==3.6.0
django-browser-reload==1.11.0
django-encrypted-model-fields==0.6.5
django-storages==1.14.2
django-cors-headers==4.2.0
django-celery-beat==2.5.0
django-ratelimit==4.1.0
dj-database-url==2.1.0

whitenoise==6.6.0

Jinja2==3.1.2

firebase-admin==6.2.0
google-cloud-storage==2.12.0

pandas
xlrd==2.0.1
openpyxl==3.1.2

beautifulsoup4==4.12.2

email-validator==2.0.0.post2

celery==5.2.7
redis==5.0.0

requests==2.31
PyJWT==2.8.0

gunicorn